+++ src/apps/agave/gnosis/agave.claimable.contract-position-fetcher.ts
@@ -1,6 +1,5 @@
 import { Inject } from '@nestjs/common';
 
-import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
 import { AaveV2ClaimableContractPositionHelper } from '~apps/aave-v2/helpers/aave-v2.claimable.contract-position-helper';
 import { PositionFetcher } from '~position/position-fetcher.interface';
@@ -8,7 +7,6 @@ import { ContractPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
 import { AGAVE_DEFINITION } from '../agave.definition';
-import { AgaveContractFactory } from '../contracts';
 
 const appId = AGAVE_DEFINITION.id;
 const groupId = AGAVE_DEFINITION.groups.claimable.id;
@@ -17,8 +15,6 @@ const network = Network.GNOSIS_MAINNET;
 @Register.ContractPositionFetcher({ appId, groupId, network })
 export class GnosisAgaveClaimableContractPositionFetcher implements PositionFetcher<ContractPosition> {
   constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(AgaveContractFactory) private readonly agaveContractFactory: AgaveContractFactory,
     @Inject(AaveV2ClaimableContractPositionHelper)
     private readonly aaveV2ClaimableContractPositionHelper: AaveV2ClaimableContractPositionHelper,
   ) {}