+++ src/apps/concentrator/ethereum/concentrator.acrv.token-fetcher.ts
@@ -15,7 +15,8 @@ import { ConcentratorContractFactory } from '../contracts';
 const appId = CONCENTRATOR_DEFINITION.id;
 const groupId = CONCENTRATOR_DEFINITION.groups.acrv.id;
 const network = Network.ETHEREUM_MAINNET;
-export const address = '0x2b95A1Dcc3D405535f9ed33c219ab38E8d7e0884'.toLowerCase(); // EIP-1967 proxy
+
+export const address = '0x2b95a1dcc3d405535f9ed33c219ab38e8d7e0884'; // EIP-1967 proxy
 
 @Register.TokenPositionFetcher({ appId, groupId, network })
 export class EthereumConcentratorAcrvTokenFetcher implements PositionFetcher<AppTokenPosition> {