+++ src/apps/sushiswap-bentobox/index.ts
@@ -1,0 +1,6 @@
+export { SushiSwapBentoBoxAppDefinition, SUSHISWAP_BENTOBOX_DEFINITION } from './sushiswap-bentobox.definition';
+export { SushiSwapBentoBoxAppModule } from './sushiswap-bentobox.module';
+export { SushiswapBentoboxContractFactory } from './contracts';
+
+export { SushiSwapBentoBoxContractPositionBalanceHelper } from './helpers/sushiswap-bentobox.vault.contract-position-balance-helper';
+export { SushiSwapBentoBoxContractPositionHelper } from './helpers/sushiswap-bentobox.vault.contract-position-helper';